수정: text directory 통합
