+++ src/components/businessPage/styles/BusinessPage.module.css
@@ -55,6 +55,9 @@
 }
 
 .box button {
+  display: flex;
+  align-items: center;
+  justify-content: center;
   color: white;
   background-color: #ff5656;
   border: none;
@@ -62,6 +65,23 @@
 }
 .box button:hover {
   background-color: #f93838;
+}
+.box button img {
+  margin-left: 10px;
+  width: 16px;
+  height: 16px;
+}
+
+.greenBox {
+  width: 100%;
+  margin-top: 15px;
+  padding: 17px 30px;
+  background-color: rgb(21, 208, 83);
+  color: white;
+  font-size: 12px;
+  line-height: 24px;
+  letter-spacing: 2.5px;
+  font-weight: 600;
 }
 
 @media screen and (max-width: 991px) {
