--- src/components/businessPage/texts/ENG.ts
@@ -1,17 +1,0 @@
-import { BusinessPageTextType } from '@/types/textTypes';
-
-const BusinessText_ENG: BusinessPageTextType = {
-  title: 'We are looking for great corporations to offer a better service for users.',
-  content:
-    "We welcome and look forward to your proposals. Fill out the form and we'll get back to you as soon as possible.",
-  address: 'Fanmaum, 916, 27, Teheran-ro-2-gil, Gangnam-gu, Seoul, Korea',
-  form: {
-    company: 'Company (required)',
-    officer: 'Full name (required)',
-    email: 'Email (required)',
-    message: 'Message',
-    button: 'SEND REQUEST',
-  },
-};
-
-export default BusinessText_ENG;
