--- src/components/businessPage/texts/KR.ts
@@ -1,17 +1,0 @@
-import { BusinessPageTextType } from '@/types/textTypes';
-
-const BusinessText_KR: BusinessPageTextType = {
-  title: '팬플러스는 다양한 제휴를 통해 더 나은 서비스를 제공하고자 합니다.',
-  content:
-    '팬플러스와 비즈니스 제휴를 희망하는 개인, 기업은 제안 내용을 등록해주시면 검토 후 연락드리겠습니다. 제안 내용은 검토 목적으로만 이용됩니다.',
-  address: '서울시 강남구 테헤란로 2길 27(역삼동), 916호',
-  form: {
-    company: '회사명 (필수)',
-    officer: '담당자명 (필수)',
-    email: '이메일 (필수)',
-    message: '문의 내용',
-    button: '문의하기',
-  },
-};
-
-export default BusinessText_KR;
