--- src/components/faqPage/texts/ENG.ts
@@ -1,33 +1,0 @@
-import { FAQPageTextType } from '@/types/textTypes';
-
-const FAQText_ENG: FAQPageTextType = {
-  header: 'We are here to help.',
-  navBar: {
-    all: 'All',
-    vote: 'Vote',
-    photos: 'Photos',
-    fanfic: 'Fanfic',
-    accounts: 'Accounts',
-  },
-  q1: 'What is FanPlus?',
-  a1: 'FanPlus is a comprehensive app for worldwide Kpop fans that provides various contents and communication features. It is the only vote app without any payment. Have more fun for your fan life on Fanplus.',
-  q2: 'How can I vote on FanPlus?',
-  a2: 'You can get cash through daily attendance or inviting friends or participate in mission on offer-walls. Exchange your earned cash for voting tickets and vote for your favorite star.',
-  q3: 'How will the reward ADs for vote winners be given?',
-  a3: 'FanPlus holds birthday/debut anniversary votes and special themed votes every month. The winner will get a reward ADs depending on the voting result, and the AD content will be made and chosen by fans. AD locations may change every month.',
-  q4: 'From when does the vote start and how do we know it?',
-  a4: "A birthday/debut anniversary vote starts about two months before the month of your bias' birthday/debut anniversary. Also, we nominate candidates from requests on FanPlus Twitter for a special themed vote. Please refer to our notice tweets related to the vote on FanPlus Twitter.",
-  twitter: 'Go to Fanplus Twitter',
-  q5: "How can I use 'Photo' feature?",
-  a5: 'Check & download photos posted by fans who are stan for the same bias as you. You can set your favorite photo as wallpaper on your mobile too.',
-  q6: "What is 'Visual Fanfic'?",
-  a6: 'Visual fanfic shows not only texts but also chat-style slides with background photos, characters, and scripts.',
-  q7: '',
-  a7: '',
-  q8: 'Is it possible to re-register on FanPlus after deactivating my account?',
-  a8: 'You can re-register on FanPlus anytime. However, you can use the same email address/nickname/mobile number 90 days after deactivation.',
-  q9: 'When can I use the same mobile number from my old account to verify?',
-  a9: 'You can use the mobile number which has been verified with the deactivated account 90 days after deactivation.',
-};
-
-export default FAQText_ENG;
