--- src/components/faqPage/texts/KR.ts
@@ -1,33 +1,0 @@
-import { FAQPageTextType } from '@/types/textTypes';
-
-const FAQText_KR: FAQPageTextType = {
-  header: '무엇을 도와드릴까요?',
-  navBar: {
-    all: '전체',
-    vote: '투표',
-    photos: '사진',
-    fanfic: '팬픽',
-    accounts: '계정',
-  },
-  q1: '팬플러스는 어떤 서비스인가요?',
-  a1: '팬플러스는 최애가 같은 팬들을 대상으로 다양한 콘텐츠와 소통 기능을 제공하는 종합 덕질 앱이에요. 유일하게 유료 결제 없는 아이돌 생일/기념일 투표를 진행하고 있답니다. 팬플러스에서 덕질에 즐거움을 더해보세요.',
-  q2: '투표는 어떻게 참여할 수 있나요?',
-  a2: '유료 결제 없이 앱 내 출석체크와 친구 초대를 통해 캐시를 획득할 수 있어요. 획득한 캐시를 투표권으로 교환하여 내 최애에게 투표해보세요.',
-  q3: '광고는 언제/어떻게 진행되나요?',
-  a3: '매달 생일/기념일 투표와 주제별 투표가 진행되고 있어요. 투표 결과에 따라 팬들이 응모한 시안 중 가장 많은 지지를 받은 시안으로 광고가 이루어져요.',
-  q4: '투표는 언제부터 준비하나요?',
-  a4: '생일에 해당하는 달로부터 약 2개월 전에 투표가 시작돼요. 주제별 투표는 트위터 공지를 통해 후보 제보를 받고 있으니, 공지를 참고해주세요.',
-  twitter: '팬플러스 트위터 바로가기',
-  q5: '사진 기능은 어떻게 사용할 수 있나요?',
-  a5: '나와 최애가 같은 팬들이 직접 올린 사진을 구경하고 다운 받아보세요. 마음에 드는 사진이 있다면 바로 배경화면으로 설정할 수도 있어요.',
-  q6: '비주얼 팬픽이 뭔가요?',
-  a6: '기존 줄글 형식의 팬픽에서 벗어나 배경/인물/대사로 구성된 장면(씬)으로 이야기가 진행되는 형식이에요.',
-  q7: '팬픽 출금 기능은 어떻게 사용하는 건가요?',
-  a7: '독자에게 응원을 받으면 팝콘 아이템이 적립돼요. 팝콘을 모아 출금 신청을 할 수 있어요. 단, 현재는 번호 인증을 완료한 한국 유저만 이용이 가능해요.',
-  q8: '팬플러스 탈퇴 후 재가입은 언제할 수 있나요?',
-  a8: '탈퇴 후 재가입은 바로 가능해요. 다만, 어뷰징 방지를 위해 탈퇴된 계정과 동일한 닉네임/계정 주소/인증된 번호로는 탈퇴일로부터 90일간 가입이 제한됩니다.',
-  q9: '탈퇴 후 같은 번호로 다시 인증하는 건 언제부터 가능한가요?',
-  a9: '탈퇴한 계정에 인증되어 있는 번호는 탈퇴일로부터 90일이 지난 이후에 다시 번호 인증에 사용할 수 있어요.',
-};
-
-export default FAQText_KR;
