+++ src/components/layout/styles/PageBox.module.css
@@ -18,12 +18,13 @@
   color: rgb(102, 102, 102);
   text-decoration: none;
   align-items: center;
-  transition: color 0.7s;
+  transition: color 0.5s;
+  font-weight: 600;
 }
 
 .pageLink:hover {
   color: rgb(0, 0, 0);
-  transition: color 0.7s;
+  transition: color 0.5s;
 }
 
 @media screen and (max-width: 768px) {
