--- src/components/layout/texts/ENG.ts
@@ -1,19 +1,0 @@
-import { NavBarTextType } from '@/types/textTypes';
-
-const NavBarText_ENG: NavBarTextType = {
-  vote: 'Vote',
-  community: 'Community',
-  aboutUs: 'About Us',
-  recruit: '',
-  business: 'Business Inquiry',
-  language: 'English',
-  link: {
-    vote: '/en/votes',
-    community: '/en/community',
-    aboutUs: '/en',
-    business: '/en/business',
-    faq: '/en/faq',
-  },
-};
-
-export default NavBarText_ENG;
