--- src/components/layout/texts/KR.ts
@@ -1,30 +1,0 @@
-import { NavBarTextType, FooterTextType } from '@/types/textTypes';
-
-const NavBarText_KR: NavBarTextType = {
-  vote: '투표',
-  community: '커뮤니티',
-  aboutUs: '서비스 소개',
-  recruit: '채용',
-  business: '제휴 문의',
-  language: '한국어',
-  link: {
-    vote: '/votes',
-    community: '/community',
-    aboutUs: '/',
-    business: '/business',
-    faq: '/faq',
-  },
-};
-export const FooterText_KR: FooterTextType = {
-  customerService: '고객센터',
-  emailButton: '이메일 문의',
-  content:
-    '팬마음 커뮤니티는 통신판매중개자이며 통신판매의 당사자가 아닙니다. 따라서 팬마음 커뮤니티는 상품 거래정보 및 거래에 대하여 책임을 지지 않습니다.',
-  termsOfService: '이용약관',
-  privacyPolicy: '개인정보처리방침',
-  introduction:
-    '대표자 : 이기현  사업자등록번호 : 695-86-00419  서울시 강남구 역삼로3길 17-6, 10층',
-  copyright: 'Copyright © 팬마음. All Rights Reserved.',
-};
-
-export default NavBarText_KR;
