--- src/components/mainPage/texts/ENG.ts
@@ -1,62 +1,0 @@
-import { MainPageTextType } from '@/types/textTypes';
-
-const MainPageText_ENG: MainPageTextType = {
-  Area1: {
-    line1: 'Must-Have App for',
-    line2: 'Worldwide Kpop Fans',
-    line3: 'FanPlus',
-  },
-  Area2: {
-    title1: 'Be together with worldwide fans',
-    title2: '& add more fun to your fan life',
-    line1: 'Fanplus considers',
-    bold_line1: ' the fun of fan life',
-    bold_line2_front: '',
-    line2: 'with other fan friends the most valuable experience',
-    bold_line2_back: '',
-    line3: 'Starting with a voting app without payment,',
-    line4: 'we are creating a community-friendly space,',
-    line5: 'so all fans have fun being together.',
-    line6: 'Meet fan friends who just hit it off with you on FanPlus',
-    line7: 'and add joy to your fan life',
-  },
-  Area3: {
-    title: '01 Promote your star',
-    line1: 'Check out votes with a variety of new topics',
-    line2: 'every month.',
-    line3: 'Participate in Fan Plus votes with no-payment',
-    line4: 'and give reward ADs to your star around the world',
-    Ad_title: '[FanPlus Reward ADs]',
-    Ad1: 'NYC Times Square in the US',
-    Ad2: 'Ningbo in China',
-    Ad3: 'Incheon Airport in Korea',
-  },
-  Area4: {
-    title1: '02 Enjoy entertaining',
-    title2: 'contents',
-    line1: 'Check high-quality photos and fanfic of your star!',
-    line2: 'Enjoy varied contents',
-    line3: 'that users create and share together.',
-    plus: "Don't miss the board too!",
-  },
-  Area5: {
-    title1: '03 How enthusiastic fan',
-    title2: 'are you?',
-    line1: 'The more active you are,',
-    line2: 'the more badges you get on FanPlus!',
-    line3: 'Show your badges to your friends',
-    line4: 'so everyone knows how much you support your star.',
-    plus1: "Check your friend's stan propensity with badges on FanPlus.",
-    plus2: '',
-  },
-  Area6: {
-    title1: '04 It is more fun',
-    title2: 'when we are all together',
-    line1: 'If you just hit it off with fan friends,',
-    line2: 'chat and share your fan life with them.',
-    line3: 'Who knows?',
-    line4: 'You will find the BFF on FanPlus!',
-  },
-};
-
-export default MainPageText_ENG;
