--- src/components/mainPage/texts/KR.ts
@@ -1,62 +1,0 @@
-import { MainPageTextType } from '@/types/textTypes';
-
-const MainPageText_KR: MainPageTextType = {
-  Area1: {
-    line1: '덕질에 즐거움 더하기,',
-    line2: '팬플러스',
-    line3: '',
-  },
-  Area2: {
-    title1: '따로 또 같이, 덕질을 즐기는 다양한 방법',
-    title2: '',
-    line1: '팬플러스는 최애가 같은 팬들끼리',
-    bold_line1: '',
-    bold_line2_front: '어떻게 하면 더 즐겁게 덕질할 수 있을지 ',
-    line2: '고민합니다.',
-    bold_line2_back: '',
-    line3: '유료 결제 없는 투표 앱으로 시작해,',
-    line4: '함께하는 덕질의 재미를 느낄 수 있도록',
-    line5: '팬플러스만의 공간을 만들어가고 있어요.',
-    line6: '나와 마음이 맞는 팬들을 만나 덕질에 즐거움을 더해보세요.',
-    line7: '',
-  },
-  Area3: {
-    title: '01 내 최애를 주인공으로',
-    line1: '매달 새롭게 열리는 다양한 주제의 투표를 확인해보세요.',
-    line2: '',
-    line3: '팬플러스만의 유료 결제 없는 투표에 참여하고,',
-    line4: '최애에게 전 세계 광고를 선물하세요.',
-    Ad_title: '[팬플러스 투표 1등 광고상품]',
-    Ad1: '미국 타임스퀘어',
-    Ad2: '중국 닝보',
-    Ad3: '인천공항',
-  },
-  Area4: {
-    title1: '02 시간 순삭 덕질 콘텐츠',
-    title2: '',
-    line1: '최애의 고화질 사진과 팬픽 감상은 덤!',
-    line2: '나와 최애가 같은 팬들이 공유하는',
-    line3: '다양한 콘텐츠를 즐겨보세요.',
-    plus: '북적이는 게시판도 방문해보세요.',
-  },
-  Area5: {
-    title1: '03 덕질, 어디까지 해봤나요?',
-    title2: '',
-    line1: '덕질할 수록 쌓이는 팬플러스만의 활동 배지!',
-    line2: '배지를 수집하는 재미와 함께',
-    line3: '최애를 향한 나의 덕력을 확인해보세요.',
-    line4: '',
-    plus1: '다른 사람의 활동 배지를 보면 그 사람의 덕질 성향도 알 수 있어요.',
-    plus2: '',
-  },
-  Area6: {
-    title1: '04 함께하면 더 즐거운 덕질',
-    title2: '',
-    line1: '나와 성향이 비슷한 친구를 발견했다면?',
-    line2: '실시간 채팅으로 최애 이야기를 나눠보세요.',
-    line3: '혹시 알아요?',
-    line4: '팬플러스에서 인생 덕메를 만나게 될지!',
-  },
-};
-
-export default MainPageText_KR;
