+++ src/pages/en/faq.tsx
@@ -1,12 +1,10 @@
-import FAQText_ENG from '@/components/faqPage/texts/ENG';
 import FAQ from '@/components/faqPage/FAQPage';
 import Layout from '@/components/layout/Layout';
-import NavBarText_ENG from '@/components/layout/texts/ENG';
-import { FooterText_KR } from '@/components/layout/texts/KR';
+import { NavBarText_ENG, FooterText_ENG, FAQText_ENG } from '@/texts/ENG';
 
 const faq = () => {
   return (
-    <Layout navBarTexts={NavBarText_ENG} footerTexts={FooterText_KR}>
+    <Layout navBarTexts={NavBarText_ENG} footerTexts={FooterText_ENG}>
       <FAQ texts={FAQText_ENG} />
     </Layout>
   );
