+++ src/texts/KOR.ts
@@ -1,0 +1,133 @@
+import {
+  NavBarTextType,
+  FooterTextType,
+  MainPageTextType,
+  BusinessPageTextType,
+  FAQPageTextType,
+} from '@/types/textTypes';
+
+export const NavBarText_KR: NavBarTextType = {
+  vote: '투표',
+  community: '커뮤니티',
+  aboutUs: '서비스 소개',
+  recruit: '채용',
+  business: '제휴 문의',
+  language: '한국어',
+  link: {
+    vote: '/votes',
+    community: '/community',
+    aboutUs: '/',
+    business: '/business',
+    faq: '/faq',
+  },
+};
+export const FooterText_KR: FooterTextType = {
+  customerService: '고객센터',
+  emailButton: '이메일 문의',
+  content:
+    '팬마음 커뮤니티는 통신판매중개자이며 통신판매의 당사자가 아닙니다. 따라서 팬마음 커뮤니티는 상품 거래정보 및 거래에 대하여 책임을 지지 않습니다.',
+  termsOfService: '이용약관',
+  privacyPolicy: '개인정보처리방침',
+  introduction: '서울시 강남구 역삼로3길 17-6, 10층',
+  copyright: 'Copyright © 팬플러스. All Rights Reserved.',
+};
+export const MainPageText_KR: MainPageTextType = {
+  Area1: {
+    line1: '덕질에 즐거움 더하기,',
+    line2: '팬플러스',
+    line3: '',
+  },
+  Area2: {
+    title1: '따로 또 같이, 덕질을 즐기는 다양한 방법',
+    title2: '',
+    line1: '팬플러스는 최애가 같은 팬들끼리',
+    bold_line1: '',
+    bold_line2_front: '어떻게 하면 더 즐겁게 덕질할 수 있을지 ',
+    line2: '고민합니다.',
+    bold_line2_back: '',
+    line3: '유료 결제 없는 투표 앱으로 시작해,',
+    line4: '함께하는 덕질의 재미를 느낄 수 있도록',
+    line5: '팬플러스만의 공간을 만들어가고 있어요.',
+    line6: '나와 마음이 맞는 팬들을 만나 덕질에 즐거움을 더해보세요.',
+    line7: '',
+  },
+  Area3: {
+    title: '01 내 최애를 주인공으로',
+    line1: '매달 새롭게 열리는 다양한 주제의 투표를 확인해보세요.',
+    line2: '',
+    line3: '팬플러스만의 유료 결제 없는 투표에 참여하고,',
+    line4: '최애에게 전 세계 광고를 선물하세요.',
+    Ad_title: '[팬플러스 투표 1등 광고상품]',
+    Ad1: '미국 타임스퀘어',
+    Ad2: '중국 닝보',
+    Ad3: '인천공항',
+  },
+  Area4: {
+    title1: '02 시간 순삭 덕질 콘텐츠',
+    title2: '',
+    line1: '최애의 고화질 사진과 팬픽 감상은 덤!',
+    line2: '나와 최애가 같은 팬들이 공유하는',
+    line3: '다양한 콘텐츠를 즐겨보세요.',
+    plus: '북적이는 게시판도 방문해보세요.',
+  },
+  Area5: {
+    title1: '03 덕질, 어디까지 해봤나요?',
+    title2: '',
+    line1: '덕질할 수록 쌓이는 팬플러스만의 활동 배지!',
+    line2: '배지를 수집하는 재미와 함께',
+    line3: '최애를 향한 나의 덕력을 확인해보세요.',
+    line4: '',
+    plus1: '다른 사람의 활동 배지를 보면 그 사람의 덕질 성향도 알 수 있어요.',
+    plus2: '',
+  },
+  Area6: {
+    title1: '04 함께하면 더 즐거운 덕질',
+    title2: '',
+    line1: '나와 성향이 비슷한 친구를 발견했다면?',
+    line2: '실시간 채팅으로 최애 이야기를 나눠보세요.',
+    line3: '혹시 알아요?',
+    line4: '팬플러스에서 인생 덕메를 만나게 될지!',
+  },
+};
+export const BusinessText_KR: BusinessPageTextType = {
+  title: '팬플러스는 다양한 제휴를 통해 더 나은 서비스를 제공하고자 합니다.',
+  content:
+    '팬플러스와 비즈니스 제휴를 희망하는 개인, 기업은 제안 내용을 등록해주시면 검토 후 연락드리겠습니다. 제안 내용은 검토 목적으로만 이용됩니다.',
+  address: '서울시 강남구 테헤란로 2길 27(역삼동), 916호',
+  form: {
+    company: '회사명 (필수)',
+    officer: '담당자명 (필수)',
+    email: '이메일 (필수)',
+    message: '문의 내용',
+    button: '문의하기',
+  },
+};
+export const FAQText_KR: FAQPageTextType = {
+  header: '무엇을 도와드릴까요?',
+  navBar: {
+    all: '전체',
+    vote: '투표',
+    photos: '사진',
+    fanfic: '팬픽',
+    accounts: '계정',
+  },
+  q1: '팬플러스는 어떤 서비스인가요?',
+  a1: '팬플러스는 최애가 같은 팬들을 대상으로 다양한 콘텐츠와 소통 기능을 제공하는 종합 덕질 앱이에요. 유일하게 유료 결제 없는 아이돌 생일/기념일 투표를 진행하고 있답니다. 팬플러스에서 덕질에 즐거움을 더해보세요.',
+  q2: '투표는 어떻게 참여할 수 있나요?',
+  a2: '유료 결제 없이 앱 내 출석체크와 친구 초대를 통해 캐시를 획득할 수 있어요. 획득한 캐시를 투표권으로 교환하여 내 최애에게 투표해보세요.',
+  q3: '광고는 언제/어떻게 진행되나요?',
+  a3: '매달 생일/기념일 투표와 주제별 투표가 진행되고 있어요. 투표 결과에 따라 팬들이 응모한 시안 중 가장 많은 지지를 받은 시안으로 광고가 이루어져요.',
+  q4: '투표는 언제부터 준비하나요?',
+  a4: '생일에 해당하는 달로부터 약 2개월 전에 투표가 시작돼요. 주제별 투표는 트위터 공지를 통해 후보 제보를 받고 있으니, 공지를 참고해주세요.',
+  twitter: '팬플러스 트위터 바로가기',
+  q5: '사진 기능은 어떻게 사용할 수 있나요?',
+  a5: '나와 최애가 같은 팬들이 직접 올린 사진을 구경하고 다운 받아보세요. 마음에 드는 사진이 있다면 바로 배경화면으로 설정할 수도 있어요.',
+  q6: '비주얼 팬픽이 뭔가요?',
+  a6: '기존 줄글 형식의 팬픽에서 벗어나 배경/인물/대사로 구성된 장면(씬)으로 이야기가 진행되는 형식이에요.',
+  q7: '팬픽 출금 기능은 어떻게 사용하는 건가요?',
+  a7: '독자에게 응원을 받으면 팝콘 아이템이 적립돼요. 팝콘을 모아 출금 신청을 할 수 있어요. 단, 현재는 번호 인증을 완료한 한국 유저만 이용이 가능해요.',
+  q8: '팬플러스 탈퇴 후 재가입은 언제할 수 있나요?',
+  a8: '탈퇴 후 재가입은 바로 가능해요. 다만, 어뷰징 방지를 위해 탈퇴된 계정과 동일한 닉네임/계정 주소/인증된 번호로는 탈퇴일로부터 90일간 가입이 제한됩니다.',
+  q9: '탈퇴 후 같은 번호로 다시 인증하는 건 언제부터 가능한가요?',
+  a9: '탈퇴한 계정에 인증되어 있는 번호는 탈퇴일로부터 90일이 지난 이후에 다시 번호 인증에 사용할 수 있어요.',
+};
